move headers to bn256 directory
