+++ CMakeLists.txt
@@ -25,9 +25,11 @@ set_property(TARGET bn256 APPEND PROPERTY
   COMPATIBLE_INTERFACE_STRING bn256_MAJOR_VERSION
 )
 
+install(DIRECTORY include/ DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
+install(DIRECTORY third-party/span/include/ DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/bn256)
+
 install(TARGETS bn256
         EXPORT bn256
-        PUBLIC_HEADER
         LIBRARY
         ARCHIVE)
