+++ include/bn256/span.h
@@ -4,7 +4,7 @@
 #   include <span>
 #else
 #   include <limits>
-#   include <span.hpp>
+#   include "nonstd/span.hpp"
 namespace std {
 template <class T, std::size_t Extent= std::numeric_limits<std::size_t>::max()>
 using span = nonstd::span<T, Extent>;
