+++ test/main_test.cpp
@@ -1,4 +1,4 @@
-#include <bn256.h>
+#include <bn256/bn256.h>
 #include <catch2/catch_test_macros.hpp>
 #include <iostream>
 #include <iomanip>
